moved browser detection to new client parser
